# src/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
      "version": "1.0.0",
      "title": "Games shop API",
      "description": "API for getting products, categories, CRUD operations with games"
  },
  "servers": [
    {
        "url": "/",
        "description": "Local Dev"
    }
  ],
...
  "consumes": [
      "application/json"
  ],
  "produces": [
      "application/json"
  ],
  "paths": {
...
      "/products/getByDisplayName": {
        "get": {
          "tags": [
            "Product"
          ],
          "summary": "Get a product by displayName",
          "operationId": "getByDisplayName",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "displayName",
              "in": "query",
              "description": "DisplayName to search a product",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "$ref": "#/definitions/Product"
              }
            },
            "400": {
              "description": "Invalid status value"
            }
          }
        }
      },
      "/products/getByMinRating": {
        "get": {
          "tags": [
            "Product"
          ],
          "summary": "Get products by minRating",
          "operationId": "getByMinRating",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "minRating",
              "in": "query",
              "description": "minRating to search products",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "$ref": "#/definitions/Products"
              }
            },
            "400": {
              "description": "Invalid status value"
            }
          }
        }
      },
      "/products/getByPrice": {
        "get": {
          "tags": [
            "Product"
          ],
          "summary": "Get products by price",
          "operationId": "getByPrice",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "price",
              "in": "query",
              "description": "You need to write the request as 10:50 (to set a price range) or 10 (to get products with the price more than 10) or :50 (to get products with the price less than 50)",
              "required": true,
              "type": "string"
            },
            {
              "name": "sortBy",
              "in": "query",
              "description": "You need to specify sort criteria and direction(asc or desc). example: totalRating:desc",
              "required": false,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "$ref": "#/definitions/Products"
              }
            },
            "400": {
              "description": "Invalid status value"
            }
          }
        }
      },
      "/products/{id}/rate": {
        "post": {
          "tags": [
            "Product"
          ],
          "summary": "Rate product",
          "operationId": "rateProductById",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Id of a product that you want to rate",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Rating object that needs to be added to the product",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Rating"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "$ref": "#/definitions/Products"
              }
            },
            "400": {
              "description": "Invalid status value"
            }
          },
          "security": [
            {
              "JWT": []
            }
          ]
        }
      }
  },
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "Product": {
      "type": "object",
      "properties": {
          "displayName": {
              "type": "string"
          },
          "categoryIds": {
              "type": "array"
          },
          "createdAt": {
              "type": "string"
          },
          "totalRating": {
              "type": "number"
          },
          "price": {
              "type": "number"
          },
          "ratings": {
              "type": "array"
          }
      }
    },
...
    "Rating": {
      "type": "object",
      "properties": {
          "username": {
              "type": "string"
          },
          "createdAt": {
              "type": "string"
          },
          "productId": {
              "type": "string"
          },
          "rating": {
              "type": "number"
          },
          "comment": {
              "type": "string"
          }
      },
      "example": {
        "rating": 5,
        "comment": "Interesting game"
      }
    },
    "Products": {
      "type": "object",
      "additionalProperties": {
          "$ref": "#/definitions/Product"
      }
    },
...
  }
}